disable metrics endpoint in encrypted builds
