+++ Makefile
@@ -15,6 +15,7 @@ LDFLAGS += -X '$(REPOSITORY_BASE_PATH)/src/utils/ota.Version=$(LATEST_TAG)'
 endif
 ifneq ($(ENCRYPTION_KEYS),)
 LDFLAGS += -X '$(REPOSITORY_BASE_PATH)/src/utils.EncryptionKeys=$(ENCRYPTION_KEYS)'
+BUILD_TAGS += encrypted
 endif
 ifneq ($(DEFAULT_CONFIG_VALUE),)
 LDFLAGS += -X '$(REPOSITORY_BASE_PATH)/src/runner/config.DefaultConfig=$(DEFAULT_CONFIG_VALUE)'
@@ -27,7 +28,7 @@ LDFLAGS += -X '$(REPOSITORY_BASE_PATH)/src/utils/metrics.BasicAuth=$(PROMETHEUS_
 endif
 
 build:
-	CGO_ENABLED=0 go build -ldflags="${LDFLAGS}" -o $(APP_NAME) -a ./main.go
+	CGO_ENABLED=0 go build -ldflags="${LDFLAGS}" -tags="${BUILD_TAGS}" -o $(APP_NAME) -a ./main.go
 
 build_encrypted: build
