+++ src/utils/metrics/prometheus.go
@@ -37,7 +37,6 @@ import (
 	"time"
 
 	"github.com/prometheus/client_golang/prometheus"
-	"github.com/prometheus/client_golang/prometheus/promhttp"
 	"github.com/prometheus/client_golang/prometheus/push"
 
 	"github.com/Arriven/db1000n/src/utils"
@@ -153,39 +152,14 @@ func ValidatePrometheusPushGateways(value string) bool {
 // ExportPrometheusMetrics starts http server and export metrics at address <ip>:9090/metrics, also pushes metrics
 // to gateways randomly
 func ExportPrometheusMetrics(ctx context.Context, gateways string) {
-	// We don't expect that rendering metrics should take a lot of time and needs long timeout
-	const timeout = 30 * time.Second
 
 	registerMetrics()
-
-	http.Handle("/metrics", promhttp.HandlerFor(
-		prometheus.DefaultGatherer,
-		promhttp.HandlerOpts{
-			// Opt into OpenMetrics to support exemplars.
-			EnableOpenMetrics: true,
-			Timeout:           timeout,
-		},
-	))
-
-	server := &http.Server{
-		Addr:    "0.0.0.0:9090",
-		Handler: nil,
-	}
-	go func(ctx context.Context, server *http.Server) {
-		<-ctx.Done()
-
-		if err := server.Shutdown(ctx); err != nil && !errors.Is(err, http.ErrServerClosed) {
-			log.Println("failure shutting down prometheus server:", err)
-		}
-	}(ctx, server)
 
 	if gateways != "" {
 		go pushMetrics(ctx, strings.Split(gateways, ","))
 	}
 
-	if err := server.ListenAndServe(); err != nil {
-		log.Fatal()
-	}
+	serveMetrics(ctx)
 }
 
 // BasicAuth client's credentials for push gateway encrypted with utils/crypto.go#EncryptionKeys[0] key
