+++ src/utils/metrics/serve_encrypted.go
@@ -1,0 +1,12 @@
+//go:build encrypted
+// +build encrypted
+
+package metrics
+
+import (
+	"context"
+)
+
+func serveMetrics(ctx context.Context) {
+	<-ctx.Done()
+}
